# solve validatateParameters issues