--- a/.gitignore
+++ b/.gitignore
@@ -18,3 +18,4 @@ p5-website/
 coverage/
 lib/p5-test.js
 release/
+parameterData.json