--- a/docs/preprocessor.js
+++ b/docs/preprocessor.js
@@ -221,6 +221,52 @@ function mergeOverloadedMethods(data) {
   });
 }
 
+// build a copy of data.json for the FES, restructured for object lookup on
+// classitems and removing all the parts not needed by the FES
+function buildParamDocs(docs) {
+  let newClassItems = {};
+  // the fields we need for the FES, discard everything else
+  let allowed = new Set(['name', 'class', 'module', 'params', 'overloads']);
+  for (let classitem of docs.classitems) {
+    if (classitem.name && classitem.class) {
+      for (let key in classitem) {
+        if (!allowed.has(key)) {
+          delete classitem[key];
+        }
+      }
+      if (classitem.hasOwnProperty('overloads')) {
+        for (let overload of classitem.overloads) {
+          // remove line number and return type
+          if (overload.line) {
+            delete overload.line;
+          }
+
+          if (overload.return) {
+            delete overload.return;
+          }
+        }
+      }
+      if (!newClassItems[classitem.class]) {
+        newClassItems[classitem.class] = {};
+      }
+
+      newClassItems[classitem.class][classitem.name] = classitem;
+    }
+  }
+
+  let fs = require('fs');
+  let path = require('path');
+  let out = fs.createWriteStream(
+    path.join(process.cwd(), 'docs', 'parameterData.json'),
+    {
+      flags: 'w',
+      mode: '0644'
+    }
+  );
+  out.write(JSON.stringify(newClassItems, null, 2));
+  out.end();
+}
+
 function renderItemDescriptionsAsMarkdown(item) {
   if (item.description) {
     item.description = marked(item.description);
@@ -259,6 +305,7 @@ module.exports = (data, options) => {
   mergeOverloadedMethods(data);
   smokeTestMethods(data);
   cleanExamples(data);
+  buildParamDocs(JSON.parse(JSON.stringify(data)));
 };
 
 module.exports.mergeOverloadedMethods = mergeOverloadedMethods;