--- a/src/core/error_helpers.js
+++ b/src/core/error_helpers.js
@@ -50,12 +50,17 @@ import { translator } from './internationalization';
 const typeColors = ['#2D7BB6', '#EE9900', '#4DB200', '#C83C00'];
 
 if (typeof IS_MINIFIED !== 'undefined') {
-  p5._validateParameters = p5._friendlyFileLoadError = p5._friendlyError = () => {};
+  p5._validateParameters = p5._friendlyFileLoadError = p5._friendlyError = p5._clearValidateParamsCache = () => {};
 } else {
   let doFriendlyWelcome = false; // TEMP until we get it all working LM
   // for parameter validation
-  const dataDoc = require('../../docs/reference/data.json');
+  const dataDoc = require('../../docs/parameterData.json');
   const arrDoc = JSON.parse(JSON.stringify(dataDoc));
+
+  const constantsReverseMap = {};
+  for (let key in constants) {
+    constantsReverseMap[constants[key]] = key;
+  }
 
   // -- Borrowed from jQuery 1.11.3 --
   const class2type = {};
@@ -84,6 +89,17 @@ if (typeof IS_MINIFIED !== 'undefined') {
   };
 
   // -- End borrow --
+
+  // for when p5 classes are specified as arguments of functions
+  // for e.g.: background(color) , where color is a p5.Color object
+  const p5Types = [];
+  const p5TypesNames = [];
+  for (let key of Object.keys(p5)) {
+    if (typeof p5[key] === 'function' && key[0] !== key[0].toLowerCase()) {
+      p5Types.push(p5[key]);
+      p5TypesNames.push(key);
+    }
+  }
 
   const friendlyWelcome = () => {
     // p5.js brand - magenta: #ED225D
@@ -248,7 +264,7 @@ if (typeof IS_MINIFIED !== 'undefined') {
   };
 
   const docCache = {};
-  const builtinTypes = [
+  const builtinTypes = new Set([
     'null',
     'number',
     'string',
@@ -257,7 +273,107 @@ if (typeof IS_MINIFIED !== 'undefined') {
     'function',
     'any',
     'integer'
-  ];
+  ]);
+
+  const basicTypes = {
+    number: true,
+    boolean: true,
+    string: true,
+    function: true,
+    undefined: true
+  };
+
+  const argumentTree = {};
+  // The following two functions are responsible for querying and inserting
+  // into the argument tree. It stores the types of arguments that each
+  // function has seen so far. It is used to query if a sequence of
+  // arguments seen in validate parameters was seen before.
+  // Lets consider the following segment of code runs repeatedly, perhaps in
+  // a loop or in draw()
+  //   color(10, 10, 10);
+  //   color(10, 10);
+  //   color('r', 'g', 'b');
+  // After the first of the segment, the argument tree looks like
+  // - color
+  //     - number
+  //        - number
+  //            - number
+  //                - seen: true
+  //            - seen: true
+  //     - string
+  //        - string
+  //            - string
+  //                - seen: true
+  // seen: true signifies that this argument was also seen as the last
+  // argument in a call. Now in the second run of the sketch, it would traverse
+  // the existing tree and see get seen: true, i.e this sequence was seen
+  // before and so scoring can be skipped. This also prevents logging multiple
+  // validation messages for the same thing.
+
+  // These two functions would be called repeatedly over and over again,
+  // so they need to be as optimized for performance as possible
+
+  const addType = (value, obj) => {
+    // check if the value is a p5 constant and if it is, we would want the
+    // value itself to be stored in the tree instead of the type
+    if (constantsReverseMap[value]) {
+      obj = obj[value] || (obj[value] = {});
+    } else {
+      let type = typeof value;
+      if (basicTypes[type]) {
+        obj = obj[type] || (obj[type] = {});
+      } else if (value === null) {
+        // typeof null -> "object". don't want that
+        obj = obj['null'] || (obj['null'] = {});
+      } else {
+        if (value.constructor && value.constructor.name) {
+          obj =
+            obj[value.constructor.name] || (obj[value.constructor.name] = {});
+          return obj;
+        }
+        let flag = true;
+        for (let i = 0, len = p5Types.length; i < len; ++i) {
+          if (value instanceof p5Types[i]) {
+            // if  the value is  p5 class
+            // for e.g p5.Vector, p5.Color etc
+            obj = obj[p5TypesKeys[i]] || (obj[p5TypesKeys[i]] = {});
+            flag = false;
+            break;
+          }
+        }
+        if (flag) {
+          obj = obj[type] || (obj[type] = {});
+        }
+      }
+    }
+    return obj;
+  };
+  const buildArgTypeCache = (func, arr) => {
+    // get the if an argument tree for current function already exists
+    let obj = argumentTree[func];
+    if (obj === undefined) {
+      // if it doesn't, create an empty tree
+      argumentTree[func] = {};
+      obj = argumentTree[func];
+    }
+
+    for (let i = 0, len = arr.length; i < len; ++i) {
+      let value = arr[i];
+      if (value instanceof Array) {
+        // an array is passed as an argument, expand it and get the type of
+        // each of its element. We distinguish the start of an array with 'as'
+        // or arraystart. This would help distinguish between the arguments
+        // (number, number, number) and (number, [number, number])
+        obj = obj['as'] || (obj['as'] = {});
+        for (let j = 0, lenA = value.length; j < lenA; ++j) {
+          obj = addType(value[j], obj);
+        }
+      } else {
+        obj = addType(value, obj);
+      }
+    }
+    return obj;
+  };
 
   // validateParameters() helper functions:
   // lookupParamDoc() for querying data.json
@@ -268,15 +384,8 @@ if (typeof IS_MINIFIED !== 'undefined') {
     const funcName = func.substr(ichDot + 1);
     const funcClass = func.substr(0, ichDot) || 'p5';
 
-    let queryResult;
-    const classitems = arrDoc.classitems;
-
-    for (const x of classitems) {
-      if (x.name === funcName && x.class === funcClass) {
-        queryResult = x;
-        break;
-      }
-    }
+    const classitems = arrDoc;
+    let queryResult = classitems[funcClass][funcName];
 
     // different JSON structure for funct with multi-format
     const overloads = [];
@@ -370,7 +479,7 @@ if (typeof IS_MINIFIED !== 'undefined') {
             lowerType = 'function';
           }
           // builtin
-          if (builtinTypes.includes(lowerType)) {
+          if (builtinTypes.has(lowerType)) {
             return { name: type, builtin: lowerType };
           }
 
@@ -627,6 +736,16 @@ if (typeof IS_MINIFIED !== 'undefined') {
     }
   };
 
+  // if a function is called with some set of wrong arguments, and then called
+  // again with the same set of arguments, the messages due to the second call
+  // will be supressed. If two tests test on the same wrong arguments, the
+  // second test won't see the validationError. clearing argumentTree solves it
+  p5._clearValidateParamsCache = function clearValidateParamsCache() {
+    for (let key of Object.keys(argumentTree)) {
+      delete argumentTree[key];
+    }
+  };
+
   /**
    * Validates parameters
    * param  {String}               func    the name of the function
@@ -649,6 +768,14 @@ if (typeof IS_MINIFIED !== 'undefined') {
       return; // skip FES
     }
 
+    // query / build the argument type tree and check if this sequence
+    // has already been seen before.
+    let obj = buildArgTypeCache(func, args);
+    if (obj.seen) {
+      return;
+    }
+    // mark this sequence as seen
+    obj.seen = true;
     // lookup the docs in the 'data.json' file
     const docs = docCache[func] || (docCache[func] = lookupParamDoc(func));
     const overloads = docs.overloads;