@stalgiag 2020-05-19
This is really exciting! There are some really big improvements here.

I believe that `preprocessor.js` is a good place to create the `parameterData.json`. What do you think @lmccart ?

@akshay-99 maybe it also makes sense to add tests for argument tree lookup to this PR? This can ensure that future changes keep these improvements in place and prevent unnecessary lookup and redundant logs.