@akshay-99 2020-05-20
> @akshay-99 maybe it also makes sense to add tests for argument tree lookup to this PR?

Yes that sounds good. I will add them soon.