@akshay-99 2020-05-21
On testing performance on this sketch
```js

let iterations = 10000000;
function setup() {

  console.time('timer');
  for(let i = 0; i<iterations; i++){
    random(10, 40);
  }
  console.timeEnd('timer');

  console.time('timer');
  for(let i = 0; i<iterations; i++){ 
    color([40, 40, 40]);
  }
  console.timeEnd('timer');

  console.time('timer');

  // drawing 10 million circles would take too long
  // hence iterations/10
  for(let i = 0; i<iterations/10; i++){ 
    circle(10, 10, 10);
  }
  console.timeEnd('timer');
}
```

The following was observed
Before:
```
timer: 1414.52099609375ms
timer: 4764.137939453125ms
timer: 3715.78515625ms
```
After: 
```
timer: 976.86376953125ms
timer: 2638.384033203125ms
timer: 3660.899169921875ms
```

Since `circle()` has only one format for the arguments, the time taken is pretty much the same as before.

In the latest commit before this, I have added more inline comments and changed some things for more simplicity and readability. @stalgiag I think this is now ready for review. Marking it as such.