@lmccart 2020-05-23
this looks good to me, @stalgiag do you want to take a look over and merge?
agree, `preprocessor.js` makes sense as the location for this.